+++ scripts/041_user_levels.sql
@@ -15,7 +15,7 @@ alter table public.profiles
   add column if not exists xp integer not null default 0,
   add column if not exists level integer not null default 1;
 
-create or replace function public.compute_level(p_xp integer)
+create or replace function public.compute_level(p_xp bigint)
 returns integer
 language sql
 immutable
